context-engineering-completed/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
streamlit
requests
pytest
nest-asyncio
python-dotenv
